feature: stripe payment

+++ hiv-and-medical-system/src/main/java/com/swp391_se1866_group2/hiv_and_medical_system/appointment/dto/response/AppointmentResponse.java
@@ -20,7 +20,7 @@ public class AppointmentResponse {
     Integer serviceId;
     String serviceName;
     String serviceType;
-    Double price;
+    Long price;
     Integer labTestSlotId;
     String doctorName;
     Integer scheduleSlotId;
